+++ publications/admin/orderedmodeladmin.py
@@ -27,41 +27,48 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 from functools import update_wrapper
 
-# from django.conf import settings
 from django.core.urlresolvers import reverse
 from django.http import HttpResponseRedirect
 from django.shortcuts import get_object_or_404
-# from django.utils.html import strip_spaces_between_tags as short
 from django.utils.translation import ugettext_lazy as _
 from django.template.loader import render_to_string
 from django.contrib import admin
-from django.contrib.admin.util import unquote
+try:
+    from django.contrib.admin.utils import unquote
+except ImportError:
+    # Django <= 1.6
+    from django.contrib.admin.util import unquote
 from django.contrib.admin.views.main import ChangeList
+from django.db.models.options import Options
 
+# Django <= 1.6
+if not getattr(Options, 'model_name', False):
+    Options.model_name = lambda self: self.module_name.lower()
+if not getattr(ChangeList, 'get_queryset', False):
+    ChangeList.get_queryset = ChangeList.get_query_set
 
 class OrderedModelAdmin(admin.ModelAdmin):
 
     def get_model_info(self):
         return dict(app=self.model._meta.app_label,
-                    model=self.model._meta.module_name)
+                    model=self.model._meta.model_name)
 
     def get_urls(self):
         try:
-            from django.conf.urls import patterns, url
+            from django.conf.urls import url
         except ImportError:
-            from django.conf.urls.defaults import patterns, url
+            from django.conf.urls.defaults import url
 
         def wrap(view):
             def wrapper(*args, **kwargs):
                 return self.admin_site.admin_view(view)(*args, **kwargs)
             return update_wrapper(wrapper, view)
-        return patterns('',
-                        url(r'^(.+)/move-(up)/$', wrap(self.move_view),
-                            name='{app}_{model}_order_up'.format(**self.get_model_info())),
-
-                        url(r'^(.+)/move-(down)/$', wrap(self.move_view),
-                            name='{app}_{model}_order_down'.format(**self.get_model_info())),
-                        ) + super(OrderedModelAdmin, self).get_urls()
+        return [
+            url(r'^(.+)/move-(up)/$', wrap(self.move_view),
+                name='{app}_{model}_order_up'.format(**self.get_model_info())),
+            url(r'^(.+)/move-(down)/$', wrap(self.move_view),
+                name='{app}_{model}_order_down'.format(**self.get_model_info())),
+        ] + super(OrderedModelAdmin, self).get_urls()
 
     def _get_changelist(self, request):
         list_display = self.get_list_display(request)
@@ -84,7 +91,7 @@ class OrderedModelAdmin(admin.ModelAdmin):
 
     def move_view(self, request, object_id, direction):
         cl = self._get_changelist(request)
-        qs = cl.get_query_set(request)
+        qs = cl.get_queryset(request)
 
         obj = get_object_or_404(self.model, pk=unquote(object_id))
         obj.move(direction, qs)
@@ -94,7 +101,7 @@ class OrderedModelAdmin(admin.ModelAdmin):
     def move_up_down_links(self, obj):
         return render_to_string("admin/publications/order_controls.html", {
             'app_label': self.model._meta.app_label,
-            'module_name': self.model._meta.module_name,
+            'module_name': self.model._meta.model_name,
             'object_id': obj.id,
             'urls': {
                 'up': reverse("admin:{app}_{model}_order_up".format(**self.get_model_info()), args=[obj.id, 'up']),