+++ publications/models/customfile.py
@@ -3,15 +3,19 @@ __author__ = 'Lucas Theis <lucas@theis.io>'
 __docformat__ = 'epytext'
 
 from django.db import models
-from publications.models import Publication
+from ..models import Publication
 
 class CustomFile(models.Model):
+
 	class Meta:
-		app_label = 'publications'
+		app_label = 'publications'  # Fix for Django<1.7
 
 	publication = models.ForeignKey(Publication)
 	description = models.CharField(max_length=256)
 	file = models.FileField(upload_to='publications/')
 
 	def __unicode__(self):
+		return self.description
+
+	def __str__(self):
 		return self.description