+++ publications/models/list.py
@@ -4,18 +4,21 @@ __docformat__ = 'epytext'
 
 from django.db import models
 
+
 class List(models.Model):
-	"""
-	Model representing a list of publications.
-	"""
+    """
+    Model representing a list of publications.
+    """
 
-	class Meta:
-		app_label = 'publications'
-		ordering = ('list',)
-		verbose_name_plural = 'Lists'
+    class Meta:
+        ordering = ('title',)
+        app_label = 'publications'  # Fix for Django<1.7
 
-	list = models.CharField(max_length=128)
-	description = models.CharField(max_length=128)
+    title = models.CharField(max_length=128, unique=True)
+    description = models.CharField(max_length=128)
 
-	def __unicode__(self):
-		return self.list
+    def __unicode__(self):
+        return self.title
+
+    def __str__(self):
+        return self.title