+++ publications/tests/__init__.py
@@ -1,0 +1,2 @@
+from .tests import Tests
+from .tests_live import LiveTests