+++ publications/views/id.py
@@ -2,28 +2,45 @@ __license__ = 'MIT License <http://www.opensource.org/licenses/mit-license.php>'
 __author__ = 'Lucas Theis <lucas@theis.io>'
 __docformat__ = 'epytext'
 
-from django.shortcuts import render_to_response
-from django.template import RequestContext
-from publications.models import Type, Publication
+from django.shortcuts import render
+
+from publications.models import Publication
+
 
 def id(request, publication_id):
-	publications = Publication.objects.filter(pk=publication_id)
+    try:
+        publication = Publication.objects.get(pk=publication_id)
 
-	if 'ascii' in request.GET:
-		return render_to_response('publications/publications.txt', {
-				'publications': publications
-			}, context_instance=RequestContext(request), content_type='text/plain; charset=UTF-8')
+        if 'plain' in request.GET:
+            return render(request, 'publications/export/publications.txt', {
+                'publications': [publication]
+            }, content_type='text/plain; charset=UTF-8')
 
-	elif 'bibtex' in request.GET:
-		return render_to_response('publications/publications.bib', {
-				'publications': publications
-			}, context_instance=RequestContext(request), content_type='text/x-bibtex; charset=UTF-8')
+        if 'bibtex' in request.GET:
+            return render(request, 'publications/export/publications.bib', {
+                'publications': [publication]
+            }, content_type='text/x-bibtex; charset=UTF-8')
 
-	else:
-		for publication in publications:
-			publication.links = publication.customlink_set.all()
-			publication.files = publication.customfile_set.all()
+        if 'mods' in request.GET:
+            return render(request, 'publications/export/publications.mods', {
+                'publications': [publication]
+            }, content_type='application/xml; charset=UTF-8')
 
-		return render_to_response('publications/id.html', {
-				'publications': publications
-			}, context_instance=RequestContext(request))
+        if 'ris' in request.GET:
+            return render(request, 'publications/export/publications.ris', {
+                'publications': [publication]
+            }, content_type='application/x-research-info-systems; charset=UTF-8')
+
+        publication.links = publication.customlink_set.all()
+        publication.files = publication.customfile_set.all()
+
+        return render(request, 'publications/pages/id.html', {
+            'publication': publication,
+            'title': publication.type,
+        })
+    except Publication.DoesNotExist:
+        return render(request, 'publications/base.html', {
+            'error': True,
+            'alert': {
+                'message': "There is no publication with this id: {}".format(publication_id)},
+        }, status=404)