+++ publications/views/list.py
@@ -2,43 +2,55 @@ __license__ = 'MIT License <http://www.opensource.org/licenses/mit-license.php>'
 __author__ = 'Lucas Theis <lucas@theis.io>'
 __docformat__ = 'epytext'
 
-from django.http import Http404
-from django.shortcuts import render_to_response
-from django.template import RequestContext
-from publications.models import List, Type, Publication
+from django.shortcuts import render
 
-def list(request, list):
-	list = List.objects.filter(list__iexact=list)
+from publications.models import List
+from publications.utils import populate
 
-	if not list:
-		raise Http404
 
-	list = list[0]
-	publications = list.publication_set.all()
-	publications = publications.order_by('-year', '-month', '-id')
+def list(request, title):
+    try:
+        publication_list = List.objects.get(title__iexact=title)
 
-	if 'ascii' in request.GET:
-		return render_to_response('publications/publications.txt', {
-				'publications': publications
-			}, context_instance=RequestContext(request), content_type='text/plain; charset=UTF-8')
+        publications = publication_list.publication_set.all()
+        publications = publications.order_by('-year', '-month', '-id')
 
-	elif 'bibtex' in request.GET:
-		return render_to_response('publications/publications.bib', {
-				'publications': publications
-			}, context_instance=RequestContext(request), content_type='text/x-bibtex; charset=UTF-8')
+        if 'plain' in request.GET:
+            return render(request, 'publications/export/publications.txt', {
+                'publications': publications
+            }, content_type='text/plain; charset=UTF-8')
 
-	elif 'rss' in request.GET:
-		return render_to_response('publications/publications.rss', {
-				'url': 'http://' + request.META['HTTP_HOST'] + request.path,
-				'publications': publications
-			}, context_instance=RequestContext(request), content_type='application/rss+xml; charset=UTF-8')
+        if 'bibtex' in request.GET:
+            return render(request, 'publications/export/publications.bib', {
+                'publications': publications
+            }, content_type='text/x-bibtex; charset=UTF-8')
 
-	else:
-		for publication in publications:
-			publication.links = publication.customlink_set.all()
-			publication.files = publication.customfile_set.all()
+        if 'mods' in request.GET:
+            return render(request, 'publications/export/publications.mods', {
+                'publications': publications
+            }, content_type='application/xml; charset=UTF-8')
 
-		return render_to_response('publications/list.html', {
-				'list': list,
-				'publications': publications
-			}, context_instance=RequestContext(request))
+        if 'ris' in request.GET:
+            return render(request, 'publications/export/publications.ris', {
+                'publications': publications
+            }, content_type='application/x-research-info-systems; charset=UTF-8')
+
+        if 'rss' in request.GET:
+            return render(request, 'publications/export/publications.rss', {
+                'url': 'http://' + request.get_host() + request.path,
+                'publications': publications
+            }, content_type='application/rss+xml; charset=UTF-8')
+
+        # load custom links and files
+        populate(publications)
+
+        return render(request, 'publications/pages/list.html', {
+            'publications': publications,
+            'title': "publications for list {}".format(title)
+        })
+    except List.DoesNotExist:
+        return render(request, 'publications/base.html', {
+            'error': True,
+            'alert': {
+                'message': "There is no publication list with this name: {}".format(title)},
+        }, status=404)